requirements.txt
# Core
requests>=2.31.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
pydantic>=2.5.0

# CLI
rich>=13.7.0
click>=8.1.7
questionary>=2.0.1

# Web Backend
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0

# AI / LLM
openai>=1.6.0
tiktoken>=0.5.2

# OSINT Modules
sherlock-project>=0.14.3
phonenumbers>=8.13.0
whois>=0.9.27
python-whois>=0.9.4
dnspython>=2.4.2
beautifulsoup4>=4.12.2
lxml>=4.9.4
Pillow>=10.1.0
exifread>=3.0.0
pycountry>=23.12.11

# Image Processing
face-recognition>=1.3.0; sys_platform != 'win32'

# Report Generation
jinja2>=3.1.2
weasyprint>=60.1
markdown>=3.5.1

# Data Visualization
networkx>=3.2.1

# Geolocation
geopy>=2.4.1
folium>=0.15.1
ipinfo>=5.0.1

# Security
cryptography>=41.0.7
certifi>=2023.11.17

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
responses>=0.24.1
